+++ src/Command/FixIP.php
@@ -61,7 +61,7 @@ class FixIP extends Command {
 
         if ($this->isCached() === false) {
             throw new \Exception(sprintf(
-                'Unsufficient data. Please run "%s init" first.',
+                'Unsufficient data. Please run "%s cache" first.',
                 $this->config['args'][0]
             ), 500);
         }